# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: npython - Python automation workflow platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /workflows:
    get:
      operationId: listWorkflows
      tags: [workflows]
      summary: List all workflows
      responses:
        "200":
          description: List of workflows
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Workflow"
    post:
      operationId: createWorkflow
      tags: [workflows]
      summary: Create a workflow
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWorkflowBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workflow"
  /workflows/{id}:
    get:
      operationId: getWorkflow
      tags: [workflows]
      summary: Get a workflow by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Workflow detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowDetail"
        "404":
          description: Not found
    put:
      operationId: updateWorkflow
      tags: [workflows]
      summary: Update a workflow
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateWorkflowBody"
      responses:
        "200":
          description: Updated workflow
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workflow"
    delete:
      operationId: deleteWorkflow
      tags: [workflows]
      summary: Delete a workflow
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
  /workflows/{id}/stats:
    get:
      operationId: getWorkflowStats
      tags: [workflows]
      summary: Get execution statistics for a workflow
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowStats"

  # ── Nodes ──────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Workflow:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        active:
          type: boolean
        tags:
          type: array
          items:
            type: string
        nodeCount:
          type: integer
        lastExecutedAt:
          type: string
          nullable: true
        lastStatus:
          type: string
          nullable: true
          enum: [pending, running, success, failed, stopped, null]
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, name, active, nodeCount, createdAt, updatedAt]
    WorkflowDetail:
      allOf:
        - $ref: "#/components/schemas/Workflow"
        - type: object
          properties:
            nodes:
              type: array
              items:
                $ref: "#/components/schemas/Node"
            edges:
              type: array
              items:
                $ref: "#/components/schemas/Edge"
    WorkflowStats:
      type: object
      properties:
        totalExecutions:
          type: integer
        successCount:
          type: integer
        failedCount:
          type: integer
        avgDurationMs:
          type: number
          nullable: true
        lastSevenDays:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
              count:
                type: integer
              successCount:
                type: integer
              failedCount:
                type: integer
            required: [date, count, successCount, failedCount]
      required: [totalExecutions, successCount, failedCount, lastSevenDays]
    CreateWorkflowBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        tags:
          type: array
          items:
            type: string
      required: [name]
    UpdateWorkflowBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        active:
          type: boolean
        tags:
          type: array
          items:
            type: string
        nodes:
          type: array
          items:
            $ref: "#/components/schemas/Node"
        edges:
          type: array
          items:
            $ref: "#/components/schemas/Edge"

    # ── Node ────────────────────────────────────────────────────────────────
    Node:
      type: object
      properties:
        id:
          type: string
        workflowId:
          type: string
        type:
          type: string
          enum: [code, set_variable, get_variable, trigger_manual, trigger_schedule, condition, http_request, wait, note]
        label:
          type: string
        positionX:
          type: number
        positionY:
          type: number
        config:
          type: object
          description: Node-type specific configuration (code, schedule, condition, etc.)
          additionalProperties: true
        retryCount:
          type: integer
          default: 0
        retryDelayMs:
          type: integer
          default: 1000
        continueOnError:
          type: boolean
          default: false
        stopOnError:
          type: boolean
          default: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, workflowId, type, label, positionX, positionY, config, createdAt, updatedAt]
    Edge:
      type: object
      properties:
        id:
          type: string
        sourceNodeId:
          type: string
        targetNodeId:
          type: string
        label:
          type: string
          nullable: true
        condition:
          type: string
          nullable: true
          enum: [always, on_success, on_error, null]
      required: [id, sourceNodeId, targetNodeId]
...